requirements.txt
absl-py==1.2.0
aicrowd-cli==0.1.15
aiobcrypt==3.2.0.post1
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==20.1.0
asgiref==3.9.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==20.3.0
audioread==3.0.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.12.2
black==24.2.0
bleach==3.3.0
blinker==1.7.0
cachetools==5.2.0
certifi==2021.5.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.7
cmake==3.27.7
colorama==0.4.4
commonmark==0.9.1
comtypes==1.2.0
contourpy==1.2.0
cvzone==1.6.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==3.0.0
Django==4.2.23
django-cleanup==5.2.0
django-cors-headers==4.7.0
django-crispy-forms==1.11.2
django-minio-storage==0.5.8
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dlib @ file:///C:/Users/DINESHKUMAR/Downloads/dlib-19.22.99-cp39-cp39-win_amd64.whl-20231129T132951Z-001/dlib-19.22.99-cp39-cp39-win_amd64.whl#sha256=0fd828405c77af2df1ff6a09964a8ba7f14838c538d9017046804f353e9b4bc2
dnspython==2.7.0
dotenv==0.9.9
ecapture==2.0.2
email_validator==2.2.0
entrypoints==0.3
exceptiongroup==1.3.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.111.0
fastapi-cli==0.0.7
feedparser==6.0.11
ffmpeg==1.4
filelock==3.13.1
Flask==1.1.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.0
fsspec==2024.2.0
future==1.0.0
gast==0.4.0
gdown==4.7.1
gevent==24.11.1
gitdb==4.0.11
GitPython==3.1.18
google-api-core==2.17.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.1.1
grpcio==1.49.0
gTTS==2.5.1
gunicorn==19.9.0
h11==0.16.0
h5py==3.10.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==2.10
imageio==2.33.0
imageio-ffmpeg==0.4.9
importlib-metadata==4.12.0
importlib_resources==6.1.3
imutils==0.5.4
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jaraco.context==4.3.0
jedi==0.18.0
Jinja2==2.10.1
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy_loader==0.4
libclang==14.0.6
librosa==0.10.1
llvmlite==0.42.0
lxml==4.6.3
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==1.1.1
matplotlib==3.8.3
mdurl==0.1.2
minio==7.2.15
mistune==0.8.4
ml-dtypes==0.2.0
more-itertools==10.2.0
motor==3.4.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
neattext==0.1.3
neo4j==5.28.2
nest-asyncio==1.5.1
networkx==3.2.1
noisereduce==3.0.2
notebook==6.2.0
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.1
onnx==1.15.0
opencv-contrib-python==4.5.5.64
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.18
packaging==23.2
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.12.1
pefile==2021.9.2
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==4.2.0
playsound==1.3.0
pooch==1.8.1
proglog==0.1.10
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==4.25.3
psutil==5.9.6
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
pycparser==2.20
pycryptodome==3.23.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydeck==0.8.1b0
pyfiglet==0.8.post1
pygame==2.5.2
Pygments==2.17.2
pyjokes==0.6.0
PyJWT==2.8.0
pymongo==4.6.1
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
PyMySQL==1.1.0
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrsistent==0.17.3
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.1
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==0.6.19
python-slugify==5.0.2
pyttsx3==2.90
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==6.0.1
pyzmq==22.1.0
qrcode==8.2
reportlab==4.4.2
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.13.2
semver==2.13.0
Send2Trash==1.5.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.0
sqlparse==0.4.1
starlette==0.37.2
streamlit==1.28.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.1
torchvision==0.17.1
tornado==6.1
tqdm==4.66.1
traitlets==5.0.5
twilio==8.10.3
typer==0.16.0
typing_extensions==4.14.1
tzdata==2023.3
tzlocal==5.2
ujson==5.10.0
ultralytics==8.1.30
urllib3==1.26.5
utils==1.0.2
uvicorn==0.30.1
validators==0.22.0
watchdog==3.0.0
watchfiles==1.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==0.15.5
whitenoise==6.9.0
wikipedia==1.4.0
winshell==0.6
wolframalpha==5.0.0
wrapt==1.14.1
XlsxWriter==1.4.3
xmltodict==0.13.0
yarl==1.9.3
zipp==3.17.0
zope.event==5.0
zope.interface==7.2
